+++ src/models/password.ts
@@ -4,7 +4,7 @@ import { QueryResultRow } from "pg";
 export class Password {
   static async registerResetRequest(email: string, token: string, expiry: number): Promise<QueryResultRow> {
     const query = `
-      INSERT INTO password_reset_request (email, token, expiry)
+      INSERT INTO auth.password_reset_request (email, token, expiry)
       VALUES ($1, $2, $3) RETURNING *
     `;
     const res = await client.query(query, [email, token, expiry]);
@@ -13,14 +13,14 @@ export class Password {
   
   static async findResetRequestByToken(resetToken: string): Promise<QueryResultRow> {
     const res = await client.query(
-      `SELECT * FROM password_reset_request WHERE token = $1`, 
+      `SELECT * FROM auth.password_reset_request WHERE token = $1`, 
     [resetToken]);
     return res.rows[0];
   }
   
   static async deleteResetRequest(email: string): Promise<void> {
     await client.query(
-      `DELETE FROM password_reset_request WHERE email = $1`, 
+      `DELETE FROM auth.password_reset_request WHERE email = $1`, 
     [email]);
   }
 }
